Move getOtherParentId to parentalLeaveUtils
