+++ libs/application/template-api-modules/src/lib/modules/templates/parental-leave/parental-leave.utils.ts
@@ -15,9 +15,9 @@ import {
   getSpouse,
   ParentalRelations,
   YES,
-  SPOUSE,
   Period as AnswerPeriod,
   getApplicationExternalData,
+  getOtherParentId,
 } from '@island.is/application/templates/parental-leave'
 
 import { apiConstants } from './constants'
@@ -74,35 +74,6 @@ export const getEmployer = (
       ? application.applicant
       : employerNationalRegistryId,
   }
-}
-
-export const getOtherParentId = (application: Application): string | null => {
-  const { familyMembers } = getApplicationExternalData(application.externalData)
-  const { otherParent, otherParentId } = getApplicationAnswers(
-    application.answers,
-  )
-
-  if (otherParent === SPOUSE) {
-    if (familyMembers === null) {
-      throw new Error(
-        'transformApplicationToParentalLeaveDTO: Cannot find spouse. Missing data for family members.',
-      )
-    }
-
-    const spouse = familyMembers.find(
-      (member) => member.familyRelation === SPOUSE,
-    )
-
-    if (!spouse) {
-      throw new Error(
-        'transformApplicationToParentalLeaveDTO: Cannot find spouse. No family member with this relation.',
-      )
-    }
-
-    return spouse.nationalId
-  }
-
-  return otherParentId
 }
 
 export const getPensionFund = (
